+++ packages/frames.js/package.json
@@ -58,6 +58,11 @@
       "require": "./dist/next/server.js",
       "import": "./dist/next/server.js",
       "types": "./dist/next/server.d.ts"
+    },
+    "./express/server": {
+      "require": "./dist/express/server.js",
+      "import": "./dist/express/server.js",
+      "types": "./dist/express/server.d.ts"
     }
   },
   "devDependencies": {
@@ -66,6 +71,7 @@
   },
   "license": "MIT",
   "peerDependencies": {
+    "@types/express": "^4.17.21",
     "next": "14.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
